+++ src/length.rs
@@ -33,6 +33,7 @@
 //! functions and enum about 'Length' octet of 'ASN.1.'
 
 use crate::{Error, StackBuffer};
+use std::cmp::Ordering;
 use std::convert::TryFrom;
 use std::mem::{size_of, size_of_val};
 use std::ops::Deref;
@@ -42,7 +43,7 @@ use std::ops::Deref;
 /// Note that `Length` represents the byte count of the contents in ASN.1.
 /// The total byte size of BER, DER, and CER is greater than that.
 /// (BER, DER, and CER are constituted of identifier, length, and contents.)
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, Copy, Hash)]
 pub enum Length {
     /// Represents 'Indefinite' length.
     ///
@@ -64,6 +65,21 @@ impl TryFrom<&[u8]> for Length {
     /// [`try_from_bytes`]: #method.from_bytes
     fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
         try_from(bytes).map(|(length, _rest)| length)
+    }
+}
+
+impl PartialEq for Length {
+    fn eq(&self, other: &Self) -> bool {
+        self.is_definite() && self.definite() == other.definite()
+    }
+}
+
+impl PartialOrd for Length {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        let this = self.definite()?;
+        let other = other.definite()?;
+
+        Some(this.cmp(&other))
     }
 }
 
@@ -185,6 +201,69 @@ impl Length {
             }
         }
     }
+
+    /// Converts `self` to `Option<usize>`.
+    ///
+    /// Returns `None` if self is `Indefinite`; otherwise, wraps the inner number in `Some`
+    /// and returns it.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use bsn1::Length;
+    ///
+    /// let length = Length::Indefinite;
+    /// assert_eq!(length.definite(), None);
+    ///
+    /// let length = Length::Definite(45);
+    /// assert_eq!(length.definite(), Some(45));
+    /// ```
+    pub fn definite(self) -> Option<usize> {
+        match self {
+            Self::Indefinite => None,
+            Self::Definite(n) => Some(n),
+        }
+    }
+
+    /// Returns `true` if `self` is `Indefinite`; otherwise `false`.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use bsn1::Length;
+    ///
+    /// let length = Length::Indefinite;
+    /// assert_eq!(length.is_indefinite(), true);
+    ///
+    /// let length = Length::Definite(12);
+    /// assert_eq!(length.is_indefinite(), false);
+    /// ```
+    pub fn is_indefinite(self) -> bool {
+        match self {
+            Self::Indefinite => true,
+            _ => false,
+        }
+    }
+
+    /// Returns `false` if `self` is `Indefinite`; otherwise `true`.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use bsn1::Length;
+    ///
+    /// let length = Length::Indefinite;
+    /// assert_eq!(length.is_definite(), false);
+    ///
+    /// let length = Length::Definite(12);
+    /// assert_eq!(length.is_definite(), true);
+    /// ```
+    pub fn is_definite(self) -> bool {
+        match self {
+            Self::Indefinite => false,
+            _ => true,
+        }
+    }
 }
 
 /// Tries to parse `bytes` starting with 'length' and returns `(Length, octets_after_length)`.
@@ -272,8 +351,10 @@ mod tests {
         // Indefinite
         {
             let bytes: &[u8] = &[0x80];
-            let res = try_from(bytes).unwrap();
-            assert_eq!((Length::Indefinite, empty), res);
+            let (length, left) = try_from(bytes).unwrap();
+
+            assert_eq!(length.is_indefinite(), true);
+            assert_eq!(left, empty);
         }
 
         // Short form
@@ -326,8 +407,10 @@ mod tests {
         {
             let mut bytes = vec![0x80];
             bytes.extend(extra);
-            let res = try_from(&bytes).unwrap();
-            assert_eq!((Length::Indefinite, extra), res);
+            let (length, left) = try_from(&bytes).unwrap();
+
+            assert_eq!(length.is_indefinite(), true);
+            assert_eq!(left, extra);
         }
 
         // Short form
@@ -464,15 +547,19 @@ mod tests {
         // Indefinite
         {
             let bytes = Length::Indefinite.to_bytes();
-            let length = try_from(&bytes).unwrap();
-            assert_eq!((Length::Indefinite, empty), length);
+            let (length, left) = try_from(&bytes).unwrap();
+
+            assert_eq!(length.is_indefinite(), true);
+            assert_eq!(left, empty);
         }
 
         // Definite
         for &len in &[0, 1, 0x7f, 0x80, 0xff, 0x0100, 0xffff, std::usize::MAX] {
             let bytes = Length::Definite(len).to_bytes();
-            let length = try_from(&bytes).unwrap();
-            assert_eq!((Length::Definite(len), empty), length);
+            let (length, left) = try_from(&bytes).unwrap();
+
+            assert_eq!(Length::Definite(len), length);
+            assert_eq!(left, empty);
         }
     }